Add logic for UseForEmptryUserAgents option
This will add a logical check for the `UseForEmptryUserAgents` being set to true and for the UA to be empty. This can be handy for both debugging quickly and for clients that don't send the UA header. In the case of clients that don't send the UA header, it can be assumed that they don't support JavaScript and would benefit from pre-render solutions.